+++ sdk/recoveryservices/arm-recoveryservices/src/models/mappers.ts
@@ -677,6 +677,34 @@ export const VaultProperties: coreClient.CompositeMapper = {
         type: {
           name: "String"
         }
+      },
+      backupStorageVersion: {
+        serializedName: "backupStorageVersion",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      monitoringSettings: {
+        serializedName: "monitoringSettings",
+        type: {
+          name: "Composite",
+          className: "MonitoringSettings"
+        }
+      },
+      redundancySettings: {
+        serializedName: "redundancySettings",
+        type: {
+          name: "Composite",
+          className: "VaultPropertiesRedundancySettings"
+        }
+      },
+      securitySettings: {
+        serializedName: "securitySettings",
+        type: {
+          name: "Composite",
+          className: "SecuritySettings"
+        }
       }
     }
   }
@@ -975,6 +1003,113 @@ export const VaultPropertiesMoveDetails: coreClient.CompositeMapper = {
       targetResourceId: {
         serializedName: "targetResourceId",
         readOnly: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const MonitoringSettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "MonitoringSettings",
+    modelProperties: {
+      azureMonitorAlertSettings: {
+        serializedName: "azureMonitorAlertSettings",
+        type: {
+          name: "Composite",
+          className: "AzureMonitorAlertSettings"
+        }
+      },
+      classicAlertSettings: {
+        serializedName: "classicAlertSettings",
+        type: {
+          name: "Composite",
+          className: "ClassicAlertSettings"
+        }
+      }
+    }
+  }
+};
+
+export const AzureMonitorAlertSettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "AzureMonitorAlertSettings",
+    modelProperties: {
+      alertsForAllJobFailures: {
+        serializedName: "alertsForAllJobFailures",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const ClassicAlertSettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ClassicAlertSettings",
+    modelProperties: {
+      alertsForCriticalOperations: {
+        serializedName: "alertsForCriticalOperations",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const VaultPropertiesRedundancySettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "VaultPropertiesRedundancySettings",
+    modelProperties: {
+      standardTierStorageRedundancy: {
+        serializedName: "standardTierStorageRedundancy",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      crossRegionRestore: {
+        serializedName: "crossRegionRestore",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const SecuritySettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "SecuritySettings",
+    modelProperties: {
+      immutabilitySettings: {
+        serializedName: "immutabilitySettings",
+        type: {
+          name: "Composite",
+          className: "ImmutabilitySettings"
+        }
+      }
+    }
+  }
+};
+
+export const ImmutabilitySettings: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ImmutabilitySettings",
+    modelProperties: {
+      state: {
+        serializedName: "state",
         type: {
           name: "String"
         }
@@ -1444,6 +1579,12 @@ export const ResourceCertificateAndAadDetails: coreClient.CompositeMapper = {
       },
       serviceResourceId: {
         serializedName: "serviceResourceId",
+        type: {
+          name: "String"
+        }
+      },
+      aadAudience: {
+        serializedName: "aadAudience",
         type: {
           name: "String"
         }